Finish core book flight

+++ IL/DTO/UserDTO.cs
@@ -2,14 +2,14 @@
 {
     public class UserDTO
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string Email { get; set; }
         public long SpaceMiles { get; set; }
         public bool IsLyMember { get; set; }
 
-        public UserDTO(int id, string firstName, string lastName, string email, long spaceMiles, bool isLyMember)
+        public UserDTO(long id, string firstName, string lastName, string email, long spaceMiles, bool isLyMember)
         {
             Id = id;
             FirstName = firstName;
